dealing with int overflow for id columns

+++ src/api_to_dimensional_model.py
@@ -1,6 +1,7 @@
 import datetime
 import pandas as pd
 import yaml
+import numpy as np
 
 def load_configs(yaml_path):
     return yaml.safe_load(yaml_path)
@@ -162,7 +163,7 @@ def create_fact(df, fact_column_processing_dict,fact_name='') -> pd.DataFrame:
         to_concat_columns = [column_key for column_key 
                              in fact_column_processing_dict.keys() 
                              if fact_column_processing_dict[column_key]["id_column"]==True]
-        df[fact_id_string] =  df[to_concat_columns].apply(lambda x: ''.join(x.astype(str)), axis=1)
+        df[fact_id_string] =  df[to_concat_columns].apply(lambda x: ''.join(x.astype(str)), axis=1).astype(np.int64)
         new_columns = [fact_id_string] + new_columns
     new_df = df[new_columns]
     return new_df
